feat(input): separate input into two files
