+++ src/shared/components/controlled-input.tsx
@@ -1,0 +1,49 @@
+import * as React from 'react';
+import type {
+  Control,
+  FieldValues,
+  Path,
+  RegisterOptions,
+} from 'react-hook-form';
+import { useController } from 'react-hook-form';
+
+import type { TxKeyPath } from '@/core';
+
+import type { NInputProps } from './input';
+import { Input } from './input';
+
+type TRule = Omit<
+  RegisterOptions,
+  'valueAsNumber' | 'valueAsDate' | 'setValueAs'
+>;
+
+export type RuleType<T> = { [name in keyof T]: TRule };
+export type InputControllerType<T extends FieldValues> = {
+  name: Path<T>;
+  control: Control<T>;
+  rules?: TRule;
+};
+
+interface ControlledInputProps<T extends FieldValues>
+  extends NInputProps,
+    InputControllerType<T> {}
+
+// only used with react-hook-form
+export function ControlledInput<T extends FieldValues>(
+  props: ControlledInputProps<T>
+) {
+  const { name, control, rules, ...inputProps } = props;
+
+  const { field, fieldState } = useController({ control, name, rules });
+  const error = fieldState.error?.message as TxKeyPath | undefined;
+  return (
+    <Input
+      ref={field.ref}
+      autoCapitalize="none"
+      onChangeText={field.onChange}
+      value={(field.value as string) || ''}
+      {...inputProps}
+      error={error}
+    />
+  );
+}
