+++ src/translations/ar/onboarding.json
@@ -1,7 +1,7 @@
 {
   "welcomeTitle": "اعثر على",
   "welcomeTitleSpan": "مهندس معماري!",
-  "welcomeDescription": "ابحث عن مهندسين معماريين لمشاريع البناء والتجديد الخاصة بك.",
+  "welcomeDescription": ".ابحث عن مهندسين معماريين لمشاريع البناء والتجديد الخاصة بك",
   "clientBtn": "اعثر على مهندس معماري",
   "architectBtn": "هل أنت مهندس معماري",
   "supplierBtn": "تسجيل دخول المورد",
