+++ src/translations/ar/resetpass.json
@@ -5,9 +5,9 @@
   "verificationCodeTitle": "رمز التحقق",
   "verificationCodeDescription": "أدخل الرمز المرسل إلى",
   "resetPasswordTitle": "إعادة تعيين كلمة المرور",
-  "resetPasswordDescription": "أنشئ كلمة مرور آمنة لحماية حسابك.",
-  "passwordConditions": "يجب أن تحتوي كلمة المرور الخاصة بك على الأقل على:",
-  "passwordLength": "يجب أن تحتوي كلمة مرورك على الأقل على:",
+  "resetPasswordDescription": ".أنشئ كلمة مرور آمنة لحماية حسابك",
+  "passwordConditions": ": يجب أن تحتوي كلمة المرور الخاصة بك على الأقل على",
+  "passwordLength": ": يجب أن تحتوي كلمة مرورك على الأقل على",
   "password-min-length": "8 أحرف",
   "password-lowercase": "حرف صغير واحد",
   "password-uppercase": "حرف كبير واحد",
